packages/test-network: add server-side eventsource, move events and reduce to shared file
